+++ mongo/datadog_checks/mongo/common.py
@@ -36,27 +36,58 @@ def get_state_name(state):
         return 'UNKNOWN'
 
 
+def get_long_state_name(state):
+    """Maps a mongod node state id to a human readable string."""
+    if state in REPLSET_MEMBER_STATES:
+        return REPLSET_MEMBER_STATES[state][1]
+    else:
+        return 'Replset state %d is unknown to the Datadog agent' % state
+
+
 class Deployment(object):
-    def get_available_metrics(self):
-        # TODO: Use this method to know what metrics to collect based on the deployment type.
+    def __init__(self):
+        self.use_shards = False
+
+    def is_principal(self):
+        """In each mongo cluster there should be always one 'principal' node. One node
+        that has full visibility on the user data and only one node should match the criteria.
+        This is different from the 'isMaster' property as a replica set primary in a shard is considered
+        as 'master' but is not 'principal' for the purpose of this integration.
+
+        This method is used to determine if the check will collect statistics on user database, collections
+        and indexes."""
         raise NotImplementedError
 
 
 class MongosDeployment(Deployment):
-    def get_available_metrics(self):
-        return None
+    def __init__(self):
+        super(MongosDeployment, self).__init__()
+        self.use_shards = True
+
+    def is_principal(self):
+        # A mongos has full visibility on the data, Datadog agents should only communicate
+        # with one mongos.
+        return True
 
 
 class ReplicaSetDeployment(Deployment):
-    def __init__(self, replset_name, replset_state):
+    def __init__(self, replset_name, replset_state, cluster_role=None):
+        super(ReplicaSetDeployment, self).__init__()
         self.replset_name = replset_name
         self.replset_state = replset_state
         self.replset_state_name = get_state_name(replset_state).lower()
+        self.use_shards = cluster_role is not None
+        self.cluster_role = cluster_role
+        self.is_primary = replset_state == 1
+        self.is_secondary = replset_state == 2
 
-    def get_available_metrics(self):
-        return None
+    def is_principal(self):
+        # There is only ever one primary node in a replica set.
+        # In case sharding is disabled, the primary can be considered the master.
+        return not self.use_shards and self.is_primary
 
 
 class StandaloneDeployment(Deployment):
-    def get_available_metrics(self):
-        return None
+    def is_principal(self):
+        # A standalone always have full visibility.
+        return True