+++ fluentd/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * [Added] Add support for additional metrics. See [#7685](https://github.com/DataDog/integrations-core/pull/7685).
 
-## 1.10.0 / 2020-09-21
+## 1.10.0 / 2020-09-21 / Agent 7.23.0
 
 * [Added] Add RequestsWrapper option to support UTF-8 for basic auth. See [#7441](https://github.com/DataDog/integrations-core/pull/7441).
 * [Fixed] Update proxy section in conf.yaml. See [#7336](https://github.com/DataDog/integrations-core/pull/7336).