+++ haproxy/CHANGELOG.md
@@ -9,7 +9,7 @@
 
 * [Added] Add new implementation to support Prometheus endpoint. See [#7620](https://github.com/DataDog/integrations-core/pull/7620).
 
-## 2.11.0 / 2020-09-21
+## 2.11.0 / 2020-09-21 / Agent 7.23.0
 
 * [Added] Add config spec. See [#7625](https://github.com/DataDog/integrations-core/pull/7625).
 * [Added] Refactor check to support new implementation. See [#7587](https://github.com/DataDog/integrations-core/pull/7587).