+++ mapreduce/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - mapreduce
 
-## 1.12.0 / 2020-09-21
+## 1.12.0 / 2020-09-21 / Agent 7.23.0
 
 * [Added] Add RequestsWrapper option to support UTF-8 for basic auth. See [#7441](https://github.com/DataDog/integrations-core/pull/7441).
 * [Added] Add mapreduce config specs. See [#7178](https://github.com/DataDog/integrations-core/pull/7178).