+++ marklogic/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - MarkLogic
 
-## 1.0.0 / 2020-09-21
+## 1.0.0 / 2020-09-21 / Agent 7.23.0
 
 * [Added] MarkLogic New Integration. See [#6524](https://github.com/DataDog/integrations-core/pull/6524).